refactor/#52 : 공통 프레임을 사용하던 워크스페이스 생성과 워크스페이스 프레임을 나눈다
