--- frontend/src/components/workspace/WorkspaceDashboard/index.tsx
@@ -1,14 +1,0 @@
-import { useEffect } from 'react';
-import currentFrameState from '@/stores/currentFrameState';
-
-const WorkspaceDashboard = () => {
-  const { setCurrentPage } = currentFrameState();
-
-  useEffect(() => {
-    setCurrentPage('workspace');
-  }, []);
-
-  return <div>WorkspaceDashboard</div>;
-};
-
-export default WorkspaceDashboard;
