+++ frontend/src/pages/activity/ActivityPage.tsx
@@ -1,0 +1,5 @@
+const ActivityPage = () => {
+  return <div>ActivityPage</div>;
+};
+
+export default ActivityPage;
